Update field view
