+++ projects/valtimo/components/src/lib/services/key-generator.service.ts
@@ -1,0 +1,37 @@
+import {Injectable} from '@angular/core';
+
+@Injectable({
+  providedIn: 'root',
+})
+export class KeyGeneratorService {
+  public getUniqueKey(entityName: string, usedKeys: string[]): string {
+    const dashCaseKey = `${entityName}`
+      .toLowerCase()
+      .replace(/[^a-z0-9-_]+|-[^a-z0-9]+/g, '-')
+      .replace(/_[-_]+/g, '_')
+      .replace(/^[^a-z]+/g, '');
+
+    if (!usedKeys.includes(dashCaseKey)) {
+      return dashCaseKey;
+    }
+
+    return this.getUniqueKeyWithNumber(dashCaseKey, usedKeys);
+  }
+
+  private getUniqueKeyWithNumber(dashCaseKey: string, usedKeys: string[]): string {
+    const numbersFromCurrentKey = (dashCaseKey.match(/^\d+|\d+\b|\d+(?=\w)/g) || []).map(
+      (numberValue: string) => +numberValue
+    );
+    const lastNumberFromCurrentKey =
+      numbersFromCurrentKey.length > 0 && numbersFromCurrentKey[numbersFromCurrentKey.length - 1];
+    const newKey = lastNumberFromCurrentKey
+      ? `${dashCaseKey.replace(`${lastNumberFromCurrentKey}`, `${lastNumberFromCurrentKey + 1}`)}`
+      : `${dashCaseKey}-1`;
+
+    if (usedKeys.includes(newKey)) {
+      return this.getUniqueKeyWithNumber(newKey, usedKeys);
+    }
+
+    return newKey;
+  }
+}
