+++ projects/valtimo/config/assets/core/en.json
@@ -755,10 +755,19 @@
         "key": "Key",
         "title": "Title",
         "titlePlaceholder": "Choose a title",
-        "type": "Type",
+        "type": "Display type",
         "typePlaceholder": "Choose type",
         "value": "Value",
         "valuePlaceholder": "Path to value"
+      },
+      "displayType": {
+        "text": "Text",
+        "boolean": "Yes/No",
+        "currency": "Currency",
+        "date": "Date",
+        "enum": "Enumeration",
+        "number": "Number",
+        "percent": "Percent"
       },
       "custom": {
         "selectComponent": "Angular component",
