+++ projects/valtimo/dossier-management/src/lib/components/dossier-management-widget-wizard/dossier-management-widget-wizard.component.scss
@@ -44,6 +44,7 @@
   }
 
   .cds--modal-container {
+    height: 800px;
     background: var(--cds-layer-02);
   }
